moving interface to api package
